infrastructure/cdk/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
pytest>=7.3.1
pytest-cov>=4.1.0 